Implemented suggestions from kaplona: removed meters as pilot unit, (re-)moved distanceService
